requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
tabulate==0.9.0
